+++ metadata.rb
@@ -5,8 +5,8 @@ license 'All rights reserved'
 description 'Installs/Configures chef-provisioner'
 long_description 'Installs/Configures chef-provisioner'
 
-version '0.1.2'
+version '0.1.6'
 
-%w( build-essential chef-dk ).each do |g|
+%w( build-essential now ).each do |g|
   depends g
 end