+++ CHANGELOG.md
@@ -1,3 +1,40 @@
+# 0.2.0
+ - removed validation key logic
+ - added 'convergence_options' hash
+ - added demo provisioning recipe (provisions provisioner)
+ - updated readme
+
+# 0.1.5
+ - removed chef-dk from from dependencies
+ - corrected knife.rb content
+
+# 0.1.4
+ - updated changelog
+ - corrected knife.rb template
+ - corrected validation logic
+ - updated readme
+
+# 0.1.3
+ - rewritten to run under chef-client
+ - ubuntu functionality
+ - shared library handling
+ - added dummy data bag
+ - added dummy provisioning recipe
+
+# 0.1.2
+ - reordered execution path
+ - added helper for chefdk run at compile
+ - removed forced execution path
+ - updated readme
+
+# 0.1.1
+ - data bag search functionality.
+ - attr cleanup.
+ - readme update
+ - attr addition
+ - provisioner driver config added
+ - general bugfixes
+
 # 0.1.0
 
 Initial release of chef-provisioner